+++ lib/util/cxxcache.js
@@ -1,6 +1,7 @@
 var Cache = require('carmen-cache').Cache;
 var queue = require('queue-async');
 var immediate = global.setImmediate || process.nextTick;
+var uniq = require('./uniq');
 
 // This file wraps the Cache object coming from C++
 //
@@ -85,25 +86,6 @@ Cache.shards = function(level, queue) {
     return shards;
 };
 
-function sorter(a, b) {
-    return a - b;
-}
-
-// Return unique elements of ids.
-Cache.uniq = function(ids) {
-    var uniq = [];
-    var last;
-    ids.sort(sorter);
-    var i = ids.length;
-    while (i--) {
-        if (ids[i] !== last) {
-            last = ids[i];
-            uniq.push(ids[i]);
-        }
-    }
-    return uniq;
-};
-
 // Store a id->data pair into this cache, computing the appropriate shard
 Cache.prototype.set = function(type, id, data) {
     var shard = Cache.shard(this.shardlevel, +id);
@@ -143,8 +125,8 @@ Cache.prototype.getall = function(getter, type, ids, callback) {
 
     q.awaitAll(function(err) {
         if (err) return callback(err);
-        var uniq = type === 'grid' ? result : Cache.uniq(result);
-        return callback(null, uniq);
+        var uniqued = type === 'grid' ? result : uniq(result);
+        return callback(null, uniqued);
     });
 
     function error(err) {
