+++ lib/util/uniq.js
@@ -1,13 +1,20 @@
 // Return unique elements.
-module.exports = function uniq(ids) {
+module.exports = uniq;
+
+function sorter(a, b) {
+    return a - b;
+}
+
+function uniq(ids) {
     var uniq = [];
     var last;
-    ids.sort();
-    for (var i = 0; i < ids.length; i++) {
+    ids.sort(sorter);
+    var i = ids.length;
+    while (i--) {
         if (ids[i] !== last) {
             last = ids[i];
             uniq.push(ids[i]);
         }
     }
     return uniq;
-};
+}
