+++ test/fixtures/mem-analyze.json
@@ -77,22 +77,22 @@
             [
                 "3304605861",
                 1155,
-                10446903967944
+                1543504072
             ],
             [
                 "4261191724",
                 457,
-                19242124574731
+                436207627
             ],
             [
                 "1479119945",
                 283,
-                10445461127206
+                3298971090982
             ],
             [
                 "807973851",
                 283,
-                10445461127206
+                3298971090982
             ],
             [
                 "1680909129",
@@ -102,27 +102,27 @@
             [
                 "3128583724",
                 128,
-                10446098661381
+                704643077
             ],
             [
                 "3034171793",
                 128,
-                10446098661381
+                704643077
             ],
             [
                 "1402060456",
                 128,
-                10446098661381
+                704643077
             ],
             [
                 "3033868875",
                 128,
-                10446098661381
+                704643077
             ],
             [
                 "1178800089",
                 128,
-                10446098661381
+                704643077
             ]
         ]
     }
